+++ admin/settings.php
@@ -5,10 +5,20 @@ require_once __DIR__.'/header.php';
 // Create a Settings object
 $rs_settings = new Settings;
 ?>
-<div class="wrapper">
+<div class="wrapper clear">
 	<?php
-	// List all settings
-	$rs_settings->listSettings();
+	// Fetch the current settings page
+	$page = $_GET['page'] ?? '';
+	
+	switch($page) {
+		case 'user_roles':
+			// User roles settings
+			$rs_settings->userRolesSettings();
+			break;
+		default:
+			// General settings
+			$rs_settings->generalSettings();
+	}
 	?>
 </div>
 <?php